split out parquet reader tests
