+++ cpp/tests/CMakeLists.txt
@@ -284,7 +284,11 @@ ConfigureTest(
   PERCENT 30
 )
 ConfigureTest(
-  PARQUET_TEST io/parquet_test.cpp io/parquet_chunked_reader_test.cpp io/parquet_common.cpp
+  PARQUET_TEST
+  io/parquet_test.cpp
+  io/parquet_chunked_reader_test.cpp
+  io/parquet_common.cpp
+  io/parquet_reader_test.cpp
   GPUS 1
   PERCENT 30
 )
